refactor: useWeek hook 분해 후 js util화
